Splicing memory allocator from executor runner to allow for code reuse for other arm applications, e.g. zephyr
